+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/logic/channel/internal/ChannelStateLogic.kt
@@ -201,7 +201,7 @@ internal interface ChannelStateLogic {
      *
      * @param hidden Boolean.
      */
-    fun setHidden(hidden: Boolean)
+    fun toggleHidden(hidden: Boolean)
 
     /**
      * Sets a replied message.